+++ brambl-sdk/src/main/scala/co/topl/brambl/dataApi/WalletStateAlgebra.scala
@@ -4,7 +4,7 @@ import cats.data.ValidatedNel
 import co.topl.brambl.builders.locks.LockTemplate
 import co.topl.brambl.models.Indices
 import co.topl.brambl.models.box.Lock
-import quivr.models.{Preimage, Proposition, VerificationKey}
+import quivr.models.{KeyPair, Preimage, Proposition}
 
 /**
  * Defines a data API for storing and retrieving wallet interaction.
@@ -16,9 +16,9 @@ trait WalletStateAlgebra[F[_]] {
    *
    * @param networkId The network id to initialize the wallet interaction with
    * @param ledgerId The ledger id to initialize the wallet interaction with
-   * @param vk The verification key to initialize the wallet interaction with
+   * @param mainKey The Topl Main verification key to initialize the wallet interaction with
    */
-  def initWalletState(networkId: Int, ledgerId: Int, vk: VerificationKey): F[Unit]
+  def initWalletState(networkId: Int, ledgerId: Int, mainKey: KeyPair): F[Unit]
 
   /**
    * Get the indices associated to a signature proposition